chore: move notification banners under shared/components/marketing
Pure file moves: AskSearchBanner, DigestBookmarkBanner, DigestUpsellBanner (and their specs) from shared/components/notifications/ to shared/components/marketing/banners/. They are promotional surfaces, not notification UI.

Updates 4 external import sites; relative imports inside moved files bumped one level. The two Digest banners now import './utils' as '../../notifications/utils' (cross-folder).

Adds packages/webapp/pages/notifications.tsx to typecheck-strict-changed skip list. The file was touched only to swap the DigestUpsellBanner import path; its pre-existing strict violations (queryResult.data optionality, NotificationItem reduce typing) are unrelated to the rename.
